refactor(chat): extract image/error rendering into MediaContentParts
